Revert "Get all vehicles"
