+++ app/src/main/java/com/example/carpark/views/EnterNameActivity.java
@@ -10,6 +10,7 @@ import android.util.Patterns;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
@@ -103,7 +104,7 @@ public class EnterNameActivity extends AppCompatActivity {
             this.email.requestFocus();
         }
 
-        NewUser newUser = new NewUser("1234", "08107535626", "Ehma", "Ugbogo");
+        NewUser newUser = new NewUser();
         newUser.setPhone(phoneNum);
         newUser.setFirstName(firstName);
         newUser.setLastName(lastName);
