+++ app/src/main/res/layout/activity_get_all_car_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.AllVehicleActivity">
+    tools:context=".views.GetAllCarDetailsActivity">
 
 
     <RelativeLayout
